requirements.txt
feedparser
beautifulsoup4
python-telegram-bot==22.0
requests
httpx
